Allow using role with different args in same play
By default, Ansible will only run a role once, no matter the number
of times it occurs as a dependency. This makes it difficult to package
the example play in this repo for more complex use. This is resolved as
detailed in http://docs.ansible.com/playbooks_roles.html#role-dependencies