Permit setting depth parameter for git module
Rationale: needed for creating a shallow clone of galaxy.
(specifically, for an optimized container image; see issue #7225).
If not set, parameter will be ignored.

From Ansible docs: "Create a shallow clone with a history truncated to
the specified number or revisions. The minimum possible value is 1,
otherwise ignored. Needs git>=1.9.1 to work correctly."