Template subdomain static dirs using an Ansible module
This Ansible role can manage subdomain static directories. The feature (introduced in PR #194, commit 4cff544) is enabled when `galaxy_manage_subdomain_static` is set to `true`.

The motivation of this commit is to improve the CI run times for usegalaxy.eu. At the moment, we are using the feature introduced in PR #198 (commit aceb4eb) that enables using symlinks to template subdomain static directories from Galaxy's static files directory (files specific to each subdomain are then merged on top of the template). It saves us a lot of disk space (since about ~154 MB are needed for each subdomain in a basic Galaxy setup from the GTN), but it two significant shortcomings: it generates very extensive logs (since each symlink needed to be checked and is logged by Ansible) and slows down playbook runs very noticeably (again, because each symlink is created by a separate Ansible task).

This commit moves the symlinking job to an Ansible module symlink_clone.py, so that only one task per subdomain is needed. Since that solves all shortcomings of the previous approach, not using symlinks is no longer an option.